optimize 1st render
Optimization is done by moving large parts of the CSS into the root
component. That component gets rendered on the server and thus the CSS
gets included directly in the head tag when delivered to the client.
Apparently this doesn't happen with the styles.css in the current setup
of the CLI project.
Also, moving stuff like the body tag styling into the styles tag of the
document directly, speeds up rendering